Stop exporting Draft sections and add dryrun option

+++ lib/bookstack/cli/export.rb
@@ -1,3 +1,5 @@
+require "digest"
+
 require "bookstack/cli/api"
 require "bookstack/cli/preprocess_bookstack_posts"
 require "bookstack/cli/transpile_to_markdown"
@@ -40,8 +42,21 @@ module Bookstack
 
           # Write it to file
           already_exists = File.exist? file_blob.file_path
-          puts "#{already_exists ? "Overwriting" : "Writing"} #{file_blob.file_path}#{options[:dryrun] ? " (dryrun)" : ""}"
-          File.write file_blob.file_path, file_blob.file_contents unless options[:dryrun]
+          dryrun_label = options[:dryrun] ? " (dryrun)" : ""
+          if already_exists
+            new_digest = Digest::SHA256.hexdigest file_blob.file_contents
+            old_digest = Digest::SHA256.hexdigest File.read file_blob.file_path
+            is_same_digest = new_digest == old_digest
+            if is_same_digest
+              puts "Skipping since unchanged #{file_blob.file_path}#{dryrun_label}"
+            else
+              puts "Overwriting #{file_blob.file_path}#{dryrun_label}"
+              File.write file_blob.file_path, file_blob.file_contents unless options[:dryrun]
+            end
+          else
+            puts "Writing #{file_blob.file_path}#{dryrun_label}"
+            File.write file_blob.file_path, file_blob.file_contents unless options[:dryrun]
+          end
         }
 
         # Write image blobs to file
